WIP, playing around with collection perms
I took a bit of time to explore the possibility of adding collection
perms to just play with the problem space and see what we're likely to
run into. For example, I'd been concerned that we'd have issues with the
NULL ids of the root collection(s!) - but in practice I think we can
work around that pretty easily.

In this commit, which could either turn into something real or just be a
throwaway exploratory thing, I:

- added a migration to populate the `collection_permissions` table,
similar to the `data_permissions` table. One weird bit is that
`collection_id` needs to be nullable - sadly - due to the fact that we
need to express whether you have permissions on the root
collection(s) (yeah, there are multiple - we have a root collection, and
a root "snippets" collection). I want to check the performance of these
migrations with large numbers of collections/groups, but hopefully this
should be less of an issue than for data permissions.

- added a way to replace calls to
`collection/permissions-set->visible-collection-ids`, which performs
poorly with large numbers of IDs, with calls to
`collection-permissions/honeysql-filter-clause`. This takes a user-id
and a "configuration" object (API isn't perfect, could use some work)
and returns a HoneySQL clause appropriate for use in a `WHERE` clause.
It supports all the same knobs as
`permissions-set->visible-collection-ids`, so you can still get
archived/unarchived/both, trash (or not), a particular trash operation
ID, or different permissions levels, etc. (Also, I tried creating 40,000
collections and running `(t2/select :model/Dashboard {:where [...]})`
and the original version takes 2.3s, while this version takes 89ms, a
26x speedup - and with a few more collections the original version would
stop working entirely as we're passing too many parameters to the query).

- used the new function *almost* everywhere we previously called
`permissions-set->visible-collection-ids`

Still todo is:

- replacing the normal calls to `mi/can-read?` and `mi/can-write?` with
equivalent checks using the `collection-permission` table. This needs
some thought too - previously these calls were "free" because they were
calculated in memory. We can do this in the new world too, of course,
but that brings us back to a world where we're loading all the
collection IDs you have permission on into memory... but I think it's
still a big improvement.

- effective children calculations still use the old method, as there are
some complex bits that I need to work into the new model.

- writes need to go to the new table instead of `:model/Permissions`

Actually as I write this, I wonder whether there's an alternative where
we read/write `:model/CollectionPermissions` to the DB, but still
calculate `*current-user-permissions-set*` (based on that) to be
equivalent, so we don't need to change existing permissions code - might
be a good path forward.